Comment Deadline Extension Request
We are working on developing a response to the TIC Documents for Public Comment due 31 January 2020. Many of our member companies are still reviewing the full suite of policies and reference architectures released last month, and given the timing of the holidays, it would be extremely beneficial for us to have the deadline extended an additional 30-45 days. 

The extension would be particularly helpful if it is announced far enough in advance of the deadline to allow for proper planning. I appreciate your consideration. 

Very respectfully, 

Alla Seiffert
Director of Cloud Policy
Internet Association